feat(codegen): v13 世代のサーバーで落ちるフィールドに既定値を入れる
miria のテストを回したら、新しい Misskey で required になったフィールドを
古いサーバーが返さず fromJson が落ちることがわかった。miria は v13 以降を
相手にするので、後から増えたものは既定値つきにする。

- Note.reactionCount
- CommunityChannel.isSensitive / isArchived / allowRenoteToExternal
- NoteChannelInfo.isSensitive / allowRenoteToExternal
- Role.preserveAssignmentOnMoveAccount / usersCount

あわせて、収集済みスキーマに載っていない sealed class のバリアントを
足せるように extra_union_variants を追加した。予約投稿の通知
(scheduledNotePosted / scheduledNotePostFailed) は 2026.4 以降のもので、
api_schemas が 2026.3.1 までしかないため生成では出てこず、放っておくと
unknown に落ちてノートを失う (miria #851 の再発)。

+++ lib/src/data/base/admin_queue_queue_stats_db_memory.dart
@@ -7,7 +7,7 @@ part 'admin_queue_queue_stats_db_memory.g.dart';
 abstract class AdminQueueQueueStatsDbMemory
     with _$AdminQueueQueueStatsDbMemory {
   const factory AdminQueueQueueStatsDbMemory({
-    required double total,
+    required int total,
     required double used,
     required double fragmentationRatio,
     required double peak,
