+++ lib/src/data/base/get_online_users_count.dart
@@ -5,7 +5,7 @@ part 'get_online_users_count.g.dart';
 
 @freezed
 abstract class GetOnlineUsersCount with _$GetOnlineUsersCount {
-  const factory GetOnlineUsersCount({required double count}) =
+  const factory GetOnlineUsersCount({required int count}) =
       _GetOnlineUsersCount;
 
   factory GetOnlineUsersCount.fromJson(Map<String, Object?> json) =>
