+++ lib/src/data/base/get_online_users_count_response.dart
@@ -5,7 +5,7 @@ part 'get_online_users_count_response.g.dart';
 
 @freezed
 abstract class GetOnlineUsersCountResponse with _$GetOnlineUsersCountResponse {
-  const factory GetOnlineUsersCountResponse({required double count}) =
+  const factory GetOnlineUsersCountResponse({required int count}) =
       _GetOnlineUsersCountResponse;
 
   factory GetOnlineUsersCountResponse.fromJson(Map<String, Object?> json) =>
