+++ lib/src/data/base/queue_metrics_meta.dart
@@ -6,7 +6,7 @@ part 'queue_metrics_meta.g.dart';
 @freezed
 abstract class QueueMetricsMeta with _$QueueMetricsMeta {
   const factory QueueMetricsMeta({
-    required double count,
+    required int count,
     required double prevTS,
     required double prevCount,
   }) = _QueueMetricsMeta;
